add:.vim/colors/flattened-D.vim fix:install.sh -L

+++ install.sh
@@ -8,7 +8,7 @@ echo "Create dotfile links..."
 for dotfile in .??*; do
     [[ $dotfile =~ $IGNORE_PATTERN ]] && continue
     if [[ $dotfile = $DIR_ONLY_CONTENTS ]] && continue
-    if [ -e "$HOME/$dotfile" ] ;then
+    if [[ -e "$HOME/$dotfile" && ! -L "$HOME/$dotfile" ]] ;then
         echo "Back up $HOME/$dotfile to $HOME/$dotfile~"
         mv "$HOME/$dotfile" "$HOME/$dotfile~"
     fi
@@ -18,7 +18,7 @@ done
 for dir in $DIR_ONLY_CONTENTS; do
     mkdir -p "$HOME/$dir"
     for content in "$dir"/*; do
-        if [ -e "$HOME/$content" ] ;then
+        if [[ -e "$HOME/$content" && ! -L "$HOME/$content" ]] ;then
             echo "Back up $HOME/$content to $HOME/$content~"
             mv "$HOME/$content" "$HOME/$content~"
         fi
